+++ Editor/Sprites/AnimationCreatorWindow.cs
@@ -13,7 +13,6 @@ namespace WallstopStudios.UnityHelpers.Editor.Sprites
     using UnityEditor;
     using UnityEngine;
     using CustomEditors;
-    using Utils;
     using WallstopStudios.UnityHelpers.Core.Animation;
     using WallstopStudios.UnityHelpers.Core.Extension;
     using WallstopStudios.UnityHelpers.Core.Helper;
@@ -2090,12 +2089,12 @@ namespace WallstopStudios.UnityHelpers.Editor.Sprites
 
             return Objects.EnumerableHashCode(EnumerateSourceHashes());
 
-            IEnumerable<(int id, string path)> EnumerateSourceHashes()
+            IEnumerable<(long id, string path)> EnumerateSourceHashes()
             {
                 for (int i = 0; i < animationSources.Count; i++)
                 {
                     Object src = animationSources[i];
-                    int id = src != null ? src.GetUnityObjectId() : 0;
+                    long id = src != null ? src.GetUnityObjectId() : 0;
                     string path = src != null ? AssetDatabase.GetAssetPath(src) : string.Empty;
                     yield return (id, path);
                 }
