+++ Tests/Runtime/Serialization/UnityEngineObjectJsonTests.cs
@@ -23,7 +23,7 @@ namespace WallstopStudios.UnityHelpers.Tests.Serialization
         public void GameObjectSerializationContainsExpectedFields()
         {
             GameObject go = Track(new GameObject("GO_Main"));
-            int expectedId = go.GetUnityObjectId();
+            long expectedId = go.GetUnityObjectId();
             string json = Serializer.JsonStringify(go);
             Assert.IsFalse(string.IsNullOrWhiteSpace(json), json);
 
@@ -35,7 +35,7 @@ namespace WallstopStudios.UnityHelpers.Tests.Serialization
             Assert.True(root.TryGetProperty("instanceId", out JsonElement id));
             Assert.AreEqual("GO_Main", name.GetString());
             StringAssert.Contains("UnityEngine.GameObject", type.GetString());
-            Assert.AreEqual(expectedId, id.GetInt32());
+            Assert.AreEqual(expectedId, id.GetInt64());
         }
 
         [Test]
@@ -52,21 +52,21 @@ namespace WallstopStudios.UnityHelpers.Tests.Serialization
         public void ObjectTypedSerializationUsesRuntimeTypeConverter()
         {
             GameObject go = Track(new GameObject("ObjectTypedGO"));
-            int expectedId = go.GetUnityObjectId();
+            long expectedId = go.GetUnityObjectId();
             object goAsObject = go;
             string json = Serializer.JsonStringify(goAsObject);
             using JsonDocument doc = JsonDocument.Parse(json);
             JsonElement root = doc.RootElement;
             Assert.AreEqual("ObjectTypedGO", root.GetProperty("name").GetString());
             StringAssert.Contains("UnityEngine.GameObject", root.GetProperty("type").GetString());
-            Assert.AreEqual(expectedId, root.GetProperty("instanceId").GetInt32());
+            Assert.AreEqual(expectedId, root.GetProperty("instanceId").GetInt64());
         }
 
         [Test]
         public void GameObjectSerializationPrettyPrintFormatsAndParses()
         {
             GameObject go = Track(new GameObject("Pretty ☃ Object ✓"));
-            int expectedId = go.GetUnityObjectId();
+            long expectedId = go.GetUnityObjectId();
             string json = Serializer.JsonStringify(go, pretty: true);
             Assert.IsFalse(string.IsNullOrWhiteSpace(json), json);
             StringAssert.Contains("\n", json);
@@ -76,21 +76,21 @@ namespace WallstopStudios.UnityHelpers.Tests.Serialization
             Assert.True(root.TryGetProperty("name", out JsonElement name));
             Assert.AreEqual("Pretty ☃ Object ✓", name.GetString());
             Assert.True(root.TryGetProperty("instanceId", out JsonElement id));
-            Assert.AreEqual(expectedId, id.GetInt32());
+            Assert.AreEqual(expectedId, id.GetInt64());
         }
 
         [Test]
         public void GameObjectSerializationStableInstanceIdAcrossSerializations()
         {
             GameObject go = Track(new GameObject("StableId"));
-            int expectedId = go.GetUnityObjectId();
+            long expectedId = go.GetUnityObjectId();
             string j1 = Serializer.JsonStringify(go);
             string j2 = Serializer.JsonStringify(go);
 
             using JsonDocument d1 = JsonDocument.Parse(j1);
             using JsonDocument d2 = JsonDocument.Parse(j2);
-            int id1 = d1.RootElement.GetProperty("instanceId").GetInt32();
-            int id2 = d2.RootElement.GetProperty("instanceId").GetInt32();
+            long id1 = d1.RootElement.GetProperty("instanceId").GetInt64();
+            long id2 = d2.RootElement.GetProperty("instanceId").GetInt64();
             Assert.AreEqual(expectedId, id1);
             Assert.AreEqual(id1, id2);
         }
@@ -114,7 +114,7 @@ namespace WallstopStudios.UnityHelpers.Tests.Serialization
         public System.Collections.IEnumerator GameObjectArraySerializationWorksWithNullAndDestroyed()
         {
             GameObject alive = Track(new GameObject("Alive"));
-            int aliveId = alive.GetUnityObjectId();
+            long aliveId = alive.GetUnityObjectId();
             GameObject dead = Track(new GameObject("Dead"));
             dead.Destroy();
             yield return null; // ensure Unity nullification
@@ -133,7 +133,7 @@ namespace WallstopStudios.UnityHelpers.Tests.Serialization
             JsonElement mid = root[1];
             Assert.AreEqual(JsonValueKind.Object, mid.ValueKind);
             Assert.AreEqual("Alive", mid.GetProperty("name").GetString());
-            Assert.AreEqual(aliveId, mid.GetProperty("instanceId").GetInt32());
+            Assert.AreEqual(aliveId, mid.GetProperty("instanceId").GetInt64());
 
             Assert.AreEqual(JsonValueKind.Null, root[2].ValueKind);
         }
@@ -142,7 +142,7 @@ namespace WallstopStudios.UnityHelpers.Tests.Serialization
         public System.Collections.IEnumerator GameObjectDictionarySerializationWorksWithNullAndDestroyed()
         {
             GameObject alive = Track(new GameObject("Alive_Dict"));
-            int aliveId = alive.GetUnityObjectId();
+            long aliveId = alive.GetUnityObjectId();
             GameObject dead = Track(new GameObject("Dead_Dict"));
             dead.Destroy();
             yield return null; // ensure Unity nullification
@@ -167,7 +167,7 @@ namespace WallstopStudios.UnityHelpers.Tests.Serialization
 
             Assert.True(root.TryGetProperty("alive", out JsonElement aliveEl));
             Assert.AreEqual("Alive_Dict", aliveEl.GetProperty("name").GetString());
-            Assert.AreEqual(aliveId, aliveEl.GetProperty("instanceId").GetInt32());
+            Assert.AreEqual(aliveId, aliveEl.GetProperty("instanceId").GetInt64());
 
             Assert.True(root.TryGetProperty("dead", out JsonElement deadEl));
             Assert.AreEqual(JsonValueKind.Null, deadEl.ValueKind);
@@ -187,7 +187,7 @@ namespace WallstopStudios.UnityHelpers.Tests.Serialization
         public void NestedObjectSerializationWithGameObjectProperty()
         {
             GameObject go = Track(new GameObject("Nested"));
-            int expectedId = go.GetUnityObjectId();
+            long expectedId = go.GetUnityObjectId();
             GoHolder holder = new() { Go = go };
             string json = Serializer.JsonStringify(holder);
             Assert.IsFalse(string.IsNullOrWhiteSpace(json), json);
@@ -196,14 +196,14 @@ namespace WallstopStudios.UnityHelpers.Tests.Serialization
             JsonElement root = doc.RootElement;
             Assert.True(root.TryGetProperty("Go", out JsonElement goEl));
             Assert.AreEqual("Nested", goEl.GetProperty("name").GetString());
-            Assert.AreEqual(expectedId, goEl.GetProperty("instanceId").GetInt32());
+            Assert.AreEqual(expectedId, goEl.GetProperty("instanceId").GetInt64());
         }
 
         [Test]
         public void NestedObjectSerializationWithGameObjectField()
         {
             GameObject go = Track(new GameObject("NestedField"));
-            int expectedId = go.GetUnityObjectId();
+            long expectedId = go.GetUnityObjectId();
             GoFieldHolder holder = new() { go = go };
             string json = Serializer.JsonStringify(holder);
             Assert.IsFalse(string.IsNullOrWhiteSpace(json), json);
@@ -212,14 +212,14 @@ namespace WallstopStudios.UnityHelpers.Tests.Serialization
             JsonElement root = doc.RootElement;
             Assert.True(root.TryGetProperty("go", out JsonElement goEl));
             Assert.AreEqual("NestedField", goEl.GetProperty("name").GetString());
-            Assert.AreEqual(expectedId, goEl.GetProperty("instanceId").GetInt32());
+            Assert.AreEqual(expectedId, goEl.GetProperty("instanceId").GetInt64());
         }
 
         [Test]
         public void FastSerializerProducesValidGameObjectJson()
         {
             GameObject go = Track(new GameObject("FastGO"));
-            int expectedId = go.GetUnityObjectId();
+            long expectedId = go.GetUnityObjectId();
             byte[] bytes = Serializer.JsonSerializeFast(go);
             Assert.NotNull(bytes);
             string json = Encoding.UTF8.GetString(bytes);
@@ -228,14 +228,14 @@ namespace WallstopStudios.UnityHelpers.Tests.Serialization
             using JsonDocument doc = JsonDocument.Parse(json);
             JsonElement root = doc.RootElement;
             Assert.AreEqual("FastGO", root.GetProperty("name").GetString());
-            Assert.AreEqual(expectedId, root.GetProperty("instanceId").GetInt32());
+            Assert.AreEqual(expectedId, root.GetProperty("instanceId").GetInt64());
         }
 
         [Test]
         public void SerializeIntoCallerBufferProducesValidGameObjectJson()
         {
             GameObject go = Track(new GameObject("BufferedGO"));
-            int expectedId = go.GetUnityObjectId();
+            long expectedId = go.GetUnityObjectId();
             byte[] buffer = null;
             int len = Serializer.JsonSerialize(
                 go,
@@ -249,7 +249,7 @@ namespace WallstopStudios.UnityHelpers.Tests.Serialization
             using JsonDocument doc = JsonDocument.Parse(json);
             JsonElement root = doc.RootElement;
             Assert.AreEqual("BufferedGO", root.GetProperty("name").GetString());
-            Assert.AreEqual(expectedId, root.GetProperty("instanceId").GetInt32());
+            Assert.AreEqual(expectedId, root.GetProperty("instanceId").GetInt64());
         }
     }
 }
